worked on the same userController and user Router
