+++ src/server/api/controllers/users-controller.js
@@ -2,9 +2,7 @@ const knex = require('../../config/db');
 
 const getUserById = async (id) => {
   try {
-    const user = await knex('users')
-      .select('user_name', 'id')
-      .where({ id });
+    const user = await knex('users').select('user_name', 'id').where({ id });
     return user;
   } catch (error) {
     return error.message;
@@ -38,7 +36,23 @@ const getFilteredUsers = async ({
   return searchUsers;
 };
 
+const createUser = async (body) => {
+  const newUser = {
+    user_name: body.userName,
+    email: body.email,
+    profile_image: body.profileImage,
+    last_seen: new Date(),
+  };
+
+  const createdUser = await knex('users').insert(newUser);
+  return {
+    successful: true,
+    id: createdUser[0],
+  };
+};
+
 module.exports = {
   getUserById,
   getFilteredUsers,
+  createUser,
 };
