+++ src/server/api/routes/api-router.js
@@ -5,7 +5,7 @@ const router = express.Router();
 // Router imports
 const modulesRouter = require('./modules.router');
 
-const usersRouter = require('./users.router');
+const usersRouter = require('./users-router');
 
 const channelsRouter = require('./channels-router');
